Fixed: `TypeError: "pocketmine\scheduler\TaskScheduler::scheduleRepeatingTask(): Argument #2 ($period) must be of type int, float given, called in D:\PocketMine-MP\plugins\RedSkyBlockX\src\RedCraftPE\RedSkyBlockX\SkyBlock.php on line 78" (EXCEPTION) in "pmsrc/src/scheduler/TaskScheduler" at line 62`

+++ src/RedCraftPE/RedSkyBlockX/SkyBlock.php
@@ -74,7 +74,7 @@ class SkyBlock extends PluginBase {
 
 		//begin autosave
 		$autosaveTimer = $this->cfg->get("Autosave Timer");
-		$ticks = round($autosaveTimer * 1200); //converts minutes to ticks
+		$ticks = (int) round($autosaveTimer * 1200); //converts minutes to ticks
 		$this->getScheduler()->scheduleRepeatingTask(new AutoSaveIslands($this), $ticks);
 
 		//register PacketHooker:
